first commit in this branch

+++ T_ppo2.py
@@ -1,9 +1,9 @@
-
-from stable_baselines import logger
+from baselines import logger
 import time
 import numpy as np
 import os
 import shutil
+from colorama import Fore
 
 
 def safemean(xs):
@@ -12,13 +12,14 @@ def safemean(xs):
 def learn(model,runner,n_epochs,n_steps,n_min_patches,n_opt_epochs,n_batch,
           clip_range,save_each,log_interval,lr):
     
-    print('PPO2 log: Training has started....')
+    print(Fore.GREEN+'PPO2 log: Training has started....'+Fore.WHITE)
             
 #    time_first_start = time.time()
     time_first_start=time.time()
     n_updates = n_epochs//n_batch
-    
+    e_time=50
     for update in range(1, n_updates+1):
+        s=time.time()
         assert n_batch % n_min_patches == 0
         
         n_batch_critic = n_batch // n_min_patches
@@ -28,26 +29,35 @@ def learn(model,runner,n_epochs,n_steps,n_min_patches,n_opt_epochs,n_batch,
         lrnow = lr(frac)
         cliprangenow = clip_range(frac)
 
-        # step 1 in the algotithm
-        # collect N transition
+        # step 1 in the algorithm
+        # collect N transactions
         img_obs, measure_obs, returns, masks, actions_list, values, neglogpacs = runner.run() #pylint: disable=E0632
 
         mblossvals = []
         
         # recurrent and non-recurrent
         inds = np.arange(n_batch)
+        
+        # step 3,4
+        # for epochs
         for opt in range(n_opt_epochs):
             np.random.shuffle(inds)
-            runner.env.reset()
             
+            
+            # get mini patches
             for start in range(0, n_batch, n_batch_critic):
                 end = start + n_batch_critic
                 mbinds = inds[start:end]
                 slices = (arr[mbinds] for arr in (img_obs, measure_obs, returns, masks, actions_list, values, neglogpacs))
-                
                 mblossvals.append(model.train(lrnow, cliprangenow, *slices))
-        
-        
+#                print('triaing')
+#                model.log(rond)
+#                rond+=1
+            
+            if time.time()>=s+e_time:
+                print(Fore.GREEN+'PPO2 log: Exceeded the full episode time'+Fore.WHITE)
+                s=time.time()
+                runner.env.reset()
         lossvals = np.mean(mblossvals, axis=0)
         tnow = time.time()
         fps = 1.0 / (tnow - tstart)
@@ -64,19 +74,18 @@ def learn(model,runner,n_epochs,n_steps,n_min_patches,n_opt_epochs,n_batch,
                 logger.logkv(lossname, lossval)
             logger.dumpkvs()
             
-#            model.log()
             
         # save model
         if save_each and (update % save_each == 0 or update == 1):
             savepath=os.path.join(os.getcwd(),'checkpoints')
-            if update==1:
+            if update==1 and os.path.isdir(savepath):
                 shutil.rmtree(savepath)
             if not os.path.exists(savepath):
                 os.mkdir(savepath)
             savepath = os.path.join(savepath, '%.5i'%update)
-            print('Saving to', savepath)
+            print('PPO2 log: Saving to', savepath)
+
             model.save(savepath)
         
         
-                
-    print('PPO2 log: Training has ended.')
+    print(Fore.Green+'PPO2 log: Training has ended.'+Fore.WHITE)
